agora-crawler-python/requirements.txt
crawlee[playwright]
supabase
python-dotenv